requirements.txt
numpy
complextorch
timm
scipy
pandas
tabulate
tqdm